requirements.txt
flask[async]
beautifulsoup4
requests
python-qbittorrent
configparser
discord.py
humanize
gunicorn
qbittorrent

# Pycord for Discord bot development
git+https://github.com/Pycord-Development/pycord#egg=py-cord

